imp: refactored recv_packet
