+++ packages/relayer-lib/src/lib.rs
@@ -6,3 +6,4 @@ pub mod chain;
 pub mod events;
 pub mod listener;
 pub mod tx_builder;
+pub mod utils;
