+++ packages/relayer-lib/src/utils/mod.rs
@@ -1,0 +1,3 @@
+//! This module contains the utilities for relayer implementations.
+
+pub mod cosmos;
